Move NuklearLogSink class to its own file
