+++ src/Engine/Graphics/CMakeLists.txt
@@ -20,6 +20,7 @@ set(ENGINE_GRAPHICS_SOURCES
         LocationFunctions.cpp
         Nuklear.cpp
         NuklearEventHandler.cpp
+        NuklearLogSink.cpp
         NuklearUtils.cpp
         Outdoor.cpp
         Overlays.cpp
@@ -54,6 +55,7 @@ set(ENGINE_GRAPHICS_HEADERS
         LocationTime.h
         Nuklear.h
         NuklearEventHandler.h
+        NuklearLogSink.h
         NuklearUtils.h
         Outdoor.h
         Overlays.h
